+++ src/core/entities/user.ts
@@ -364,3 +364,15 @@ export interface UserDetailInfoByUin {
         vipNameColorId: string
     }
 }
+export enum UserDetailSource {
+    KDB,
+    KSERVER
+}
+
+export enum ProfileBizType {
+    KALL,
+    KBASEEXTEND,
+    KVAS,
+    KQZONE,
+    KOTHER
+}
